requirements.txt
streamlit
requests
python-docx
PyPDF2